Dziala odbieranie danych z czujnka, dodano plik Config.h, wprowadzono tryb uspienia

+++ ATtiny_ext_station_code/external_module/src/BME280.cpp
@@ -1,8 +1,9 @@
 #include <avr/io.h>
 #include <util/delay.h>
 
+#include "Config.h"
 #include "BME280.h"
-#include "TWI.h"
+#include "SoftwareI2C.h"
 
 void BME280_WriteReg(uint8_t reg, uint8_t value)
 {
@@ -63,33 +64,33 @@ void BME280_ReadCalibration(void)
     //kalibracja 0x88..0xA1 (26 bajtów)
     BME280_ReadBytes(0x88, buf1, 26);
 
-    dig_T1 = (uint16_t)((buf1[1] << 8) | buf1[0]);
-    dig_T2 = (int16_t)((buf1[3] << 8) | buf1[2]);
-    dig_T3 = (int16_t)((buf1[5] << 8) | buf1[4]);
+    dig_T1 = (uint16_t)(((uint16_t)buf1[1] << 8) | buf1[0]);
+    dig_T2 = (int16_t)(((uint16_t)buf1[3] << 8) | buf1[2]);
+    dig_T3 = (int16_t)(((uint16_t)buf1[5] << 8) | buf1[4]);
 
-    dig_P1 = (uint16_t)((buf1[7] << 8) | buf1[6]);
-    dig_P2 = (int16_t)((buf1[9] << 8) | buf1[8]);
-    dig_P3 = (int16_t)((buf1[11] << 8) | buf1[10]);
-    dig_P4 = (int16_t)((buf1[13] << 8) | buf1[12]);
-    dig_P5 = (int16_t)((buf1[15] << 8) | buf1[14]);
-    dig_P6 = (int16_t)((buf1[17] << 8) | buf1[16]);
-    dig_P7 = (int16_t)((buf1[19] << 8) | buf1[18]);
-    dig_P8 = (int16_t)((buf1[21] << 8) | buf1[20]);
-    dig_P9 = (int16_t)((buf1[23] << 8) | buf1[22]);
+    dig_P1 = (uint16_t)(((uint16_t)buf1[7] << 8) | buf1[6]);
+    dig_P2 = (int16_t)(((uint16_t)buf1[9] << 8) | buf1[8]);
+    dig_P3 = (int16_t)(((uint16_t)buf1[11] << 8) | buf1[10]);
+    dig_P4 = (int16_t)(((uint16_t)buf1[13] << 8) | buf1[12]);
+    dig_P5 = (int16_t)(((uint16_t)buf1[15] << 8) | buf1[14]);
+    dig_P6 = (int16_t)(((uint16_t)buf1[17] << 8) | buf1[16]);
+    dig_P7 = (int16_t)(((uint16_t)buf1[19] << 8) | buf1[18]);
+    dig_P8 = (int16_t)(((uint16_t)buf1[21] << 8) | buf1[20]);
+    dig_P9 = (int16_t)(((uint16_t)buf1[23] << 8) | buf1[22]);
 
     dig_H1 = buf1[25]; //0xA1
 
     //kalibracja 0xE1..0xE7 (7 bajtów)
     BME280_ReadBytes(0xE1, buf2, 7);
 
-    dig_H2 = (int16_t)((buf2[1] << 8) | buf2[0]);
+    dig_H2 = (int16_t)(((uint16_t)buf2[1] << 8) | buf2[0]);
     dig_H3 = buf2[2];
 
     //H4 i H5 są zapisane "przycinając" bajty:
     //H4 = buf2[3] << 4 | (buf2[4] & 0x0F)
     //H5 = buf2[5] << 4 | (buf2[4] >> 4)
-    dig_H4 = (int16_t)((buf2[3] << 4) | (buf2[4] & 0x0F));
-    dig_H5 = (int16_t)((buf2[5] << 4) | ((buf2[4] >> 4) & 0x0F));
+    dig_H4 = (int16_t)(((int16_t)buf2[3] << 4) | (buf2[4] & 0x0F));
+    dig_H5 = (int16_t)(((int16_t)buf2[5] << 4) | ((buf2[4] >> 4) & 0x0F));
 
     dig_H6 = (int8_t)buf2[6];
 }
@@ -137,12 +138,14 @@ uint32_t BME280_Compensate_H(int32_t adc_H)
 {
     int32_t v_x1_u32r;
 
-    v_x1_u32r = t_fine - 76800;
-    v_x1_u32r = (((((adc_H << 14) - ((int32_t)dig_H4 << 20) - ((int32_t)dig_H5 * v_x1_u32r)) + 16384) >> 15) *
-                 (((((v_x1_u32r * (int32_t)dig_H6) >> 10) * (((v_x1_u32r * (int32_t)dig_H3) >> 11) + 32768)) >> 10) + 2097152) *
-                  (int32_t)dig_H2 + 8192) >> 14;
-
-    v_x1_u32r = v_x1_u32r - (((((v_x1_u32r >> 15) * (v_x1_u32r >> 15)) >> 7) * (int32_t)dig_H1) >> 4);
+    // Używamy sufiksu 'L' przy stałych (np. 76800L), żeby wymusić arytmetykę 32-bit
+    v_x1_u32r = (t_fine - ((int32_t)76800L));
+    
+    v_x1_u32r = (((((adc_H << 14) - (((int32_t)dig_H4) << 20) - (((int32_t)dig_H5) * v_x1_u32r)) + 
+    ((int32_t)16384L)) >> 15) * (((((((v_x1_u32r * ((int32_t)dig_H6)) >> 10) * (((v_x1_u32r * ((int32_t)dig_H3)) >> 11) + ((int32_t)32768L))) >> 10) + ((int32_t)2097152L)) * ((int32_t)dig_H2) + 8192) >> 14));
+    
+    v_x1_u32r = (v_x1_u32r - (((((v_x1_u32r >> 15) * (v_x1_u32r >> 15)) >> 7) * ((int32_t)dig_H1)) >> 4));
+    
     if (v_x1_u32r < 0) v_x1_u32r = 0;
     if (v_x1_u32r > 419430400) v_x1_u32r = 419430400;
